src/backend/microservices/data_login/requirements.txt
# Pendiente de añadir las versiones
pyarrow
pandas
fastapi
uvicorn
python-multipart
scikit-learn
pydantic
numpy
redis
uuid
mysql.connector
python-dotenv
sqlalchemy